-fixed warnings in player class by adding static casts -Added some tga files to be used -Added description to level selection class -Added a display overview image to level selection -edited the quit button image to sync with other buttons
Problem:
endl wont work on text print to screen
'\n' , "\n", "\r\n" wont work either

+++ SP4/Base/Source/Player/Player.cpp
@@ -366,7 +366,7 @@ void CPlayer::Update(double dt, CTilemap* tile)
 		{
 			if (offSetDirectionX)
 			{
-				m_fOffSetX += dt * ENTITY_MOVE_SPEED;
+				m_fOffSetX += static_cast<float>(dt) * ENTITY_MOVE_SPEED;
 				if (m_fOffSetX > tile->GetTileSize())
 				{
 					m_fOffSetX = 0;
@@ -376,7 +376,7 @@ void CPlayer::Update(double dt, CTilemap* tile)
 			}
 			else
 			{
-				m_fOffSetX -= dt * ENTITY_MOVE_SPEED;
+				m_fOffSetX -= static_cast<float>(dt)* ENTITY_MOVE_SPEED;
 				if (m_fOffSetX < -tile->GetTileSize())
 				{
 					m_fOffSetX = 0;
